requirements.txt
openai==1.43.0
python-dotenv
colorama==0.4.6

